now it works

+++ src/jmg/exp/nathann/CountK22_Result.java
@@ -2,7 +2,7 @@ package jmg.exp.nathann;
 
 public class CountK22_Result
 {
-	public long nk22 = -1;
+	public long nk22 = - 1;
 	public long twiceNK22 = 0;
 	public long nbK22pot;
 
@@ -12,7 +12,7 @@ public class CountK22_Result
 		String s = "";
 		s += " - nbK22=" + nk22;
 		s += "\n - nbK22pot=" + nbK22pot;
-		s += "\n - CK=" + (4 * nk22 / (double) nbK22pot);
+		s += "\n - CK=" + (4d * nk22 / nbK22pot);
 		return s;
 	}
 }
